Menu screen can be toggled by tapping fps counter.

+++ LittleRulerAndroid/app/src/main/cpp/dalbaragi/s_input_queue.cpp
@@ -13,11 +13,11 @@ namespace dal {
 		mCurIndex = 0;
 	}
 
-	unsigned int ISingleUsageQueue::getSize(void) const {
+	size_t ISingleUsageQueue::getSize(void) const {
 		return mCurIndex;
 	}
 
-	unsigned int ISingleUsageQueue::getCapacity(void) const {
+	size_t ISingleUsageQueue::getCapacity(void) const {
 		return kCapacity;
 	}
